Flatten the resources page module
