--- src/pages/resources/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(test)]
-mod data;
-mod page;
-
-pub use page::Resources;
